[logdog] Fix viewer bug.
The LogDog viewer currently encounters a "race" where it doesn't
properly update its controls at the end of a stream. This is because
rendering is asynchronous with fetching, and so the last update
oftentimes lands on the second-to-last control state instead of the
final one.

This removes the RENDERING state. This state was always poorly
implemented, as it was a subset of LOADING state and occurred asynchronously
with it. Best case the viewer sees the viewer rapidly switching between
LOADING and RENDERING, which is not very useful. Now, we use LOADING
state for all stream loading.

BUG=None
TEST=local
R=hinoka@chromium.org, iannucci@chromium.org

Change-Id: I2eae48718bedcc523a76dc034b05a7b66b716852
Reviewed-on: https://chromium-review.googlesource.com/684654
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>