[logdog] Decrease log level for common errors.
Several errors are common, and should not be logged at "error" level.
Logging at "error" level pollutes the application's logs and makes
filtering less useful.

"User denied READ access": happens commonly when the LogDog viewer makes
a request prior to actually logging in.

"Log stream does not exist": This is an error condition, but nothing
went wrong in the app. This happens especially when there is a lag
between a log stream's emission and registration.

BUG=None
TEST=None

Change-Id: I295848e49d18cb7426b373bd3e8c776286547d9c