Update patch set 1
Patch Set 1:

In particular; this introduces explicit errors where we had silent errors before. I've added panics in some functions which don't return errors, but I'm not happy with that, and will probably be adding error return values here.

Note that the previous timestamp logic could introduce silent errors when manipulating timestamp protos. This is particularly concerning in the auth code, as it suggests there might be some way to construct a token with nonsense timestamps (but I haven't confirmed that this is the case, I think all the code generating tokens is 'trusted' code and doesn't involve user input, but I'm not sure).

However, I need some help making sense of proper error handling strategies for a lot of this code, so if you could take a quick look at the relevant bits of code that would be much appreciated.

Patch-set: 1
Reviewer: Robbie Iannucci <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>