[milo] add buildstore package
The problems the CL addresses:

1. "go.chromium.org/luci/milo/api/buildbot".Build type is used for 
two purposes currently: json encoding and datastore entity.
This CL moves the datastore entity responsibility to a separate
type, leaving the Build in the api dir only for json encoding.
The new "go.chromium.org/luci/milo/buildsource/buildbot/buildstore".buildEntity
type has a single responsibility: save/load
"go.chromium.org/luci/milo/api/buildbot".Build to/from datastore.

2. Currently "go.chromium.org/luci/milo/buildsource/buildbot" package
does many things, where one of them datastore access. Ideally, LUCI
emulation should be implemented at the build storage level, so
move the datastore access to a separate package (buildstore). Next 
CL will implement emulation in the buildstore package.

3. Before importing a build, pubsub handling code checks if the build
in the datastore is finished, and if it is, rejects the pubsub message.
However, this was done not in a transaction.

This CL is a relatively large refactoring and splitting it up
will not necessarily simplify review because it'd require the
reviewer to consider a lot of intermediate states as opposed to
the a single final state (new package) and updates to the existing
code (use new package).

Bug:
Change-Id: Ie7060b7aee75c5cb328ccf840d7680e5f45e0374