Update patch set 1
Patch Set 1:

PTAL

This CL is a relatively large refactoring and splitting it up
will not necessarily simplify review because it'd require the
reviewer to consider a lot of intermediate states as opposed to
the a single final state (new package) and updates to the existing
code (usage of the new package).

Tests currently panic due to a bug in datastore. I will be fixing the bug while this CL is being reviewed.

Patch-set: 1
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false