Update patch set 7
Patch Set 7:

(1 comment)

> Patch Set 6:
> 
> (1 comment)
> 
> looks nice :)
> 
> Let's also whitelist domains.  IE Don't return a favicon unless the host matches storage.googleapis.com.

I actually whitelisted on the config update side, and didn't add the favicon URL to the datastore if it failed the check. I'm not sure if this is exactly desired, but I felt that this was something that fit in with config validation? If you think it should happen on the render side I can change it though.

Patch-set: 7
Reviewer: Michael Knyszek <1154295@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>