[luci-notify] use buildbucket package
buildbucket package provides Build struct which is a convenient
representation of Build message. Use that instead of custom type.
This reduces the need for luci-notify's own Build type.
The only member of luci-notify's Build type not provided by
buildbucket's Build and used in >1 place is BuilderID method.
A single member is a probably not enough justification for a new type,
so move BuilderID to a separate function.

Change Builder entity properties to be more type-safe: time.Time instead
of int64, buildbucket.Status instead of string.
This is a breaking change, but the properties are renamed too
so the worst it can cause is skipping one notification per builder.
Also update other int64 times to time.Time

Run tests in parallel.

Bug:
Change-Id: Ic466e7877e4c88330f2bc77bb5006dc8dd472b88
Reviewed-on: https://chromium-review.googlesource.com/700855
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>