LogDog: Default instance is AppEngine Classic.
At the moment, Flex enviornment is not capable of meeting the
expectations of several LUCI modules. This change creates a new LogDog
service, "logs", and moves the Flex Logs endpoint to this service.

The default service is replaced with a basic stub. This stub hosts the
standard LUCI endpoint set that is expected by LUCI services and
middleware. Because it is a Standard environment, it is capable of
serving "/_ah/..." pages, which AppEngine authentication (notably the
/admin/settings endpoint) relies on.

BUG=chromium:753846
TEST=dev

Change-Id: I5e8a78e91fb2f231240f2becbf7034715dedc140