[luci-notify] Use commits for "previous" build.
This alters how luci-notify determines a "previous" build for the
on_change notification. If commit information is available for a
given build from its buildset, then luci-notify will start to use
that (the combination of commit time and revision) to track changes.
If at any point that information is unavailable, it will fall back
to build creation time as before.

Change-Id: I72666f8e3cb718cff303f99d5a8d9f67f5660dc9