[logdog] Enable Collector RPC batching.
Update the Collector to enable RPC batching.

We do this by implementing a "bundleServicesClient", a client to the
Coordinator's Services endpoint that automatically bundles RPCs using a
Bundler.

This will significantly reduce the number of distinct RPCs that are sent
to the Coordinator backend.

BUG=chromium:771845
TEST=None

Change-Id: I3942490d7379bf9543b946c8eb4a5cd9d8966e0c
Reviewed-on: https://chromium-review.googlesource.com/702159
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>