Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> would be nice to have something like this, not specific to logdog
> maybe a tool accepts service def(s) and generates a batch service proto, go interfaces and go impl.

One problem with generalizing batching is size constraints. It happens to be the case that the Services endpoint's requests and responses are relatively proportional in size, so they won't exceed GAE's request/response threshold. However, if you tried to do this to (say) LogDog's Logs API, you'd probably exceed thresholds all the time, and the logic to not do that would be nontrivial.

One could say, "respond to as many as you can before you hit limit", but then you can't execute the in parallel unless they're idempotent. If they are, you'd have to note which indices finished and include that in the response, and prune entries on the client side and resubmit until the full batch is complete. It's do-able, but it wouldn't be nearly as clean as this implementation.

Patch-set: 5