Update patch set 7
Patch Set 7:

I'm still not convinced it has to be in the manifest. See my comment for Patch Set 2.

In any case, I don't like hardcoding default set of platforms in the binary at all. If users what to use verify feature, they must provide the list of platforms themselves. I don't want us to rebuild and redeploy cipd client everywhere if we add or remove a platform.

Patch-set: 7
CC: Vadim Shtayura <1003293@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>