[vpython] Use native CIPD verifier.
Now that CIPD has native verification, update the "verify" subcommand to
use CIPD verification instead of implementing its own. This enables
concurrent verification!

Add a formal "Verify" PackageLoader method. This is needed so that the
PackageLoader can take care of all verification scenarios at once in an
implementation-specific way (in CIPD's case, concurrently) instead of
feeding it scenario after scenario. This also has the nice side-effect
of greatly simplifying the verification interface.

This is not API-compatible with "infra/tools/vpython"; that code will
need to be updated.

BUG=None
TEST=local

Change-Id: I2f845cbd40815f25f76428e03fc2371c8a88146a
Reviewed-on: https://chromium-review.googlesource.com/706809
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>