[milo] Add patch_* fields to Manifest.GitCheckout.
The way that bot_update (and presumably other checkout tools) actually
apply patches is that they fetch both the current 'master' revision from
the repo, then the patch (i.e. from the changes refs), then rebase the
patch on top of the master. This is done so that tryjobs are always
testing against the most recent commits.

This CL updates the Manifest and ManifestDiff messages so that they can
accurately report these patches. Both the base (master) revision as well
as the actual patch revision information can be expressed.

The Manifest message is backwards compatible with existing messages 
(binary + json). The ManifestDiff hasn't been released in the wild
yet.

R=hinoka@chromium.org, nodir@chromium.org

Bug:
Change-Id: Iac865eaf48f51429cbbfde7834ac30bc4be2ce82