Update patch set 2
Patch Set 2:

said another way, I think Milo should visualize the truth of the system, not re-invent/interpret it. I'm not convinced that visualizing every timeout as an infra failure is good... timeouts can happen for a variety of reasons:
  * too many jobs scheduled
  * not enough resources/bad dimensions
  * etc.

IMO, a timeout is a timeout; it's an expected outcome of the system. It happens enough to have its own status (which is why buildbucket helpfully returns it as a status code, instead of just calling it infra failure). Infra failures are generally an unknown/unclassified error.

I think it's unhelpful to take classified errors and rebrand them as unclassified errors...

Patch-set: 2
Reviewer: Robbie Iannucci <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>