[logdog] Dispatch bundle request in parallel.
Currently, bundle requests are dispatched in series, with the bundler
waiting for the request to complete prior to dispatching further
requests. Have the bundler client dispatch in parallel so a single RPC,
even bundled, doesn't bottleneck the service.

This will increase concurrency, but it will still be bounded by the
maximum consumer rate. This should increase services' throughput to the
rate they were processing prior to bundling.

BUG=chromium:771845
TEST=None
R=iannucci@chromium.org, nodir@chromium.org

Change-Id: I573610da7f212e75b3e4971d4a84a24457197b6e