[vpython] Better failure output.
When "vpython" fails during process invocation, it will generally not
output anything. This is because the failure has an exit code, and so
"vpython" handles it as a process failure and yields that exit code
without logging.

This is incorrect behavior for two reasons:
- "vpython" should only be forwarding process return code when that
  process is the final Python interpreter invocation.
- Furthermore, now that we use "execve"-style semantics, "vpython"
  should not even be forwarding the exit code at all. The launch function
  is responsible for doing this.

Update the code to remove that special-case code path.

Additionally, update VirtualEnv setup to buffer process output when not
immediately sending it to STDERR and dump that output on error.
Previously, process output is forwarded directly to STDERR only when the
current logging level indicates verbosity. However, this prevents the
user from seeing process output in the event of an error. This change
displays the process output when an error occurs.

Because we default to error-level logging and only log errors when there
is a failure, we still maintain the mandatory quality of no additional
logs by default.

BUG=None
TEST=local
  - Ran with error-level logging, observed a "virtualenv.py" execution
    error.
  - Ran without errors, obaserved no output.

R=iannucci@chromium.org, nodir@chromium.org

Change-Id: I5a74be4a3f8bc646f58006216b463730feee8c8c
Reviewed-on: https://chromium-review.googlesource.com/724479
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>