[buildbucket] parsing build address: support partial responses
A partial buildbucket response message may not contain bucket and ParseMessage
thinks that build_address tag is invalid. Change that so it does not try to
compare first build_address component to an empty string of the bucket, and
instead treats that as the bucket name.

Bug:
Change-Id: Icc10875e35e95cb48b0807f4f375bbfa13cc1677
Reviewed-on: https://chromium-review.googlesource.com/747002
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>