Reland: [milo] Use BuildSummary for rendering buildbucket builds.
This avoids the extra RPC to buildbucket when the BuildSummary
contains a swarming://host/task/task_id ContextURI. However, if it fails
to find this on BuildSummary, it will fall back to the extra RPC before
failing.

This includes a deprecation note for the fallback codepath to be removed
after ~1 month.

R=hinoka@chromium.org, jchinlee@chromium.org, nodir@chromium.org

Bug: 758007
Change-Id: Id09fe5951388d8817174982bb60d9e8ef8b0c3b7