Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1 Commit-Queue+2

offline we discussed this:

updating builder summary is not transactional now, but acceptable as a very short solution to fix the outage.

I believe we should remove consoles from Builder summary because we don't need them and because they are computed incorrectly currently (updated on build ingestion as opposed to config ingestion): console handler has the builder ids from the console def and can make a single datastore.get_multi RPC to get BuilderSummaries. This is also faster than running a query like we do now (because the query is not a projection query)

Also storing pending builds in BuilderSummary does not seem to be sustainable the way it is implemented right now. We should at the very least update them only when a build status changes, not every time we receive a pubsub message about a any build of the builder (this is too often).

Patch-set: 3