Add Run() function delegates to avoid command execution in tests
It's worthwhile to have some more integration-y test that actually makes
sure that we're executing real commands - in our case, the "executes
the command" tests - but we probably don't want *all* of our tests to
be hitting the real command interfaces.

This will be useful in testing the upcoming addition of the drain file
logic, as well.

Bug: 416072
Change-Id: Ief64123ce8a763e92f7874e616006c320b55fca7