Update patch set 2
Patch Set 2:

Note that this will require some manual deployment, as it touches datastore schema in a non-backwards-compatible way. However, the window should be short.

Basically:
  * upload+switch
  * delete all Console entity keys
  * re-import Console entities

If this is rolled back, the process would be the same (but just when switched to the old version)

Patch-set: 2
Reviewer: Robbie Iannucci <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>