[vpython] fix subprocess output dumping
when running a subprocess, vpython redirects its output to own one
if the logging level is high enough. Otherwise it buffers output and logs it
only on non-zero exit code. However, dumping of the buffer is done with the same
logging level (which is not logged), so this is noop.

On failure, log the buffered output with error level.

Bug: 783892
Change-Id: I37612209f158925f27f8af82fc2e717f868dfdbe
Reviewed-on: https://chromium-review.googlesource.com/765292
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>