[vpython] fix subprocess output dumping
when running subprocesses, vpython redirects their output to its own if
specified logging level is high enough. Otherwise it buffers output and logs it
only on non-zero exit code. However, dumping of buffer uses the same logging
level (which is not logged), so this is noop.

On failure, log the buffered output with level Error.

Bug: 783892
Change-Id: I96adde389093cd6fd65c3b6715f94bc429f64c8d