Update patch set 2
Patch Set 2:

(1 comment)

Hm, I looked again... I guess this is probably OK, but it still doesn't make sense to me. The pubsub event won't be ACK'd until both entities are updated, which means the transactions are still coupled (but now they're coupled in the code instead of the datastore). I think if you keep the early aborts (e.g. "if the build isn't terminal, then don't load the builder summary, and if the times are out of order, then don't Put the summary), and also kept the xg transaction, it would be the best of both worlds.

Patch-set: 2
Reviewer: Robbie Iannucci <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>