[tsmon] Refactor appengine/tsmon in preparation for splitting it.
There are 3 functional components in this package currently:
  1) Middleware that manages tsmon state and captures request metrics.
  2) A housekeeping cron that assigns task numbers to instances.
  3) Global metrics (also reported from the housekeeping cron).

(1) has very little actual ties to GAE and can be extracted into a package
reusable from Flex or GKE or wherever. This CL is preparation for this.

It introduces TaskNumAllocator interface that formalizes relation between
the tsmon middleware and the housekeeping cron. GAE standard and GAE Flex will
have slightly different implementations of TaskNumAllocator.

R=tandrii@chromium.org
BUG=785049

Change-Id: Ibf9074b940166f4af5aa61e06ce9b8b77bd8d402