[vpython] Fix spec loading from environment.
At one point, vpython used to only store a Spec in a VirtualEnv's stamp
file. This changed, and now it stores an Environment. However, code in
the startup to load a Spec passed down by a "vpython" instance via
environment variable was not updated, and will fail thinking that it is
loading a Spec instead of an Environment.

Update that code, and generalize Environment loading similar to how Spec
loading is constructed.

Bug: 784597
Change-Id: I1ce80667aef30f942c086731d229428f8cbc279e
Reviewed-on: https://chromium-review.googlesource.com/778062
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>