[milo] Provide builders view by relative time.
This is analogous to Buildbot's builders view, displaying, for each
builder, the number of pending and running builds, and the statuses
of the last {limit} builds on that builder. Thus builders cannot be
correlated time-wise with each other in this view.

This CL does not filter the builders list based on console, instead
displaying all builders in a given project; this will be changed in
(one of) the next CL(s).

Bug: 775935
Change-Id: I289a03efefe8e326f240cf76f46cd7032bd32c5f
Reviewed-on: https://chromium-review.googlesource.com/777706
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>