Add drain file support within RunExclusive()
This patch makes it so that RunExclusive() creates a drain file while it
waits for the lock and deletes that drain file as soon as it acquires
the lock.

A future patch will make it so that RunShared() looks for this drain
file before attempting to acquire a shared lock on the lock file. This
guarantees that swarming jobs don't continuously run tasks with at least
one of them holding the shared lock while puppet starves, awaiting
the lock to become completely free.

Bug: 416072
Change-Id: Ib04a8c8dbf2030b349579bc3dc38cc86c067b09e