[milo] initialize header on demand
milo.frontend.console initializes a header struct unconditionally.
This might cause a console page to display a header even if one is not defined.

Also if a console defines a header and there was no error, the header variable
will be initialized, so we can simply check if header is not nil

Bug:
Change-Id: I108333b90b657b185168e8aa9b372c8805ed8124