[milo] initialize header only if it is defined
milo.frontend.console initializes a header struct unconditionally.
This might cause a console page to display a header even if one is not defined.

Also if a console DOES define a header and there was no error, the header variable
will be initialized, so we can simply check if it is not nil

Bug:
Change-Id: I108333b90b657b185168e8aa9b372c8805ed8124
Reviewed-on: https://chromium-review.googlesource.com/787170
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>