Add drain file support to RunShared()
This patch makes it so that RunShared() blocks on the presence of a
drain file before requesting the global file lock.

The guarantees that swarming jobs that are continuously running tasks
won't starve puppet, which requires an exclusive lock to run because
it can't be run alongside other swarming tasks.

Bug: 416072
Change-Id: Id9e7ada5779127b6056476fcbcb5018a6e420eb6
Reviewed-on: https://chromium-review.googlesource.com/780059
Commit-Queue: Charlie Andrews <charliea@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>