Update patch set 4
Patch Set 4:

(copypasta comment since i didn't send it out correctly the first time)

This is missing the cookie feature, but I think it's ready for review without it.  The cookie feature can be added in a separate CL.

Demo:
https://2408-b64efee-tainted-hinoka-dot-luci-milo-dev.appspot.com/p/chromium/g/main/console#

Some notes:

All javascript is in the .html file because I think the functions are so specific to the layout of that specific page it doesn't make sense to break it out.
I've checked for a while, and I'm fairly certain there is no way to do this natively in CSS without refactoring the structure of the HTML and switching to using html tables, just like how buildbot did it.  I'm pretty sure we don't want to do that (the layout was significantly slower using tables).

Patch-set: 4
Reviewer: Ryan Tseng <1002080@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>