Fix two flaky time-based tests in mmutex package
Both of these tests assumed that context-based deadlines will be
enforced for the *exact* amount of time that they're specified. For
various reasons, including the overhead of the previous lines of
execution, it's likely that the time will differ slightly, so just
checking with the context expired is likely to be a lot less
error-prone.

Bug: 790881
Change-Id: I142c568c175a90c8f5245c493d0ebcf25cc6eb4e
Reviewed-on: https://chromium-review.googlesource.com/806744
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>