Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > (1 comment)
> 
> Good point but your suggestion doesn't follow the convention set by builder and console view pages -> "<project> | <view>". What I am suggesting is "<site> | <view>" which follows high-level with view.

The point is that people usually only see the first word or two of the title in the tab and it's important to have something that disambiguates which tab it is. The favicon already tells people what site they have opened so displaying LUCI isn't as helpful as the specific view they're on. See gmail, monorail, or gerrit for examples.

Patch-set: 1
CC: Erik Staab <1128941@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>