Remove the drain file when the exclusive command can't acquire a lock
Before this, the drain file was only being removed upon the successful
acquisition of the lock. This would cause all shared tasks to lock until
an exclusive task could be run, even if that exclusive task had timed
out and was no longer waiting to run.

This CL modifies the behavior so that, if the exclusive task gives up
on running and times out, it cleans up the drain file, allowing shared
tasks to run.

Bug: 416072
Change-Id: Ia277f14e2d862b9baa9a3211b45ea2dc18eb75df
Reviewed-on: https://chromium-review.googlesource.com/817355
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Charlie Andrews <charliea@chromium.org>