Update patch set 2
Patch Set 2:

(13 comments)

Made the changes so that the validation context will now contain as much information about the errors as possible (before it just returned on the first error that it found).
This meant piecing through validation.Context through many places, which is why this CL just got so much bigger :/

Nevertheless, happy Monday.
P.S.
Also, I've noticed that a lot of the things are named ...Acl... insteaad of ...ACL...
I can make a separate CL that will address those after this is landed since  the purpose of this CL is to get the validation endpoint up and running for scheduler.

Patch-set: 2
Reviewer: Mun Yong Jang <1246322@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>