[isolate] Make number of concurent uploads configurable.
In exparchive mode, the number of concurrent uploads is currently capped to 1,
in an attempt to avoid disk thrashing.  However, this is not necessarily the
optimal setting, due to e.g. overhead in setting up HTTP connections for
uploads. Initial experiments indicate that a higher number of concurrent
uploads is superior.  This CL adds a flag so that different values can be
tested in prod.

Bug: 794425
Change-Id: Id96a29ab4a7df7cd1ffe5f1d444ecce46bc5b150
Reviewed-on: https://chromium-review.googlesource.com/826526
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Michael McGreevy <mcgreevy@chromium.org>