[milo] make recipes link useful
currently recipes link is not useful: it takes the user to repo root and the
user has to click around to find the recipe, or completely disregard the page
by going to code search.

Use code search instead. In many cases, it takse the user directly to the recipe
source code, which is the ideal.

Bug:
Change-Id: Idaa3308b11d03b70fdd1f80eadef79ce0f746627