[prpc] remove response.go
response struct is not a well-defined abstraction and complicates adding more
logging of HTTP>=400 responses. Replace it with writeError and writeMessage
functions that write directly to the given http.ResponseWriter.

Rename server.respond to server.call and make it return response message,
response format and error.

Use status.Errorf instead of grpc.Errorf in places where we change code
anyway.

Bug:
Change-Id: Ifd2775f30d9bdf558cd2ad7fbb889561fd253a6f