[vpython] Partial revert of 9fe9988df59aa2ae0d83861aa733bf5f1e20f3f4.
This:
  * Bumps venv environment format to v3
  * Includes the hash of sitecustomize.py into the environment hash.
  * Removes sys.executable override from sitecustomize.py.
  * Makes vpython strip all VirtualEnv's from $PATH from the environ
    (including non-vpython ones).
  * Puts the virtualenv bin directory back in $PATH (but in such a way
    that only one can be present in $PATH at any given time, to
    minimize accumulation effects).

This is because we have scripts in the chromium ecosystem which
intentionally do:
  * Tweak $PYTHONPATH
  * Launch subprocess (using sys.executable)

The new philosophy is that `vpython` indicates an explicit, hard break
in the python environment. "wrapper" type scripts which wrap other,
unknown, scripts will need to explicitly use vpython. "launcher" type
scripts (e.g. the ones that do the tweak+launch) are much more
prevalent and after this CL will be able to work without modification.

R=dnj@chromium.org, vadimsh@chromium.org

Bug: 793325, 795638
Change-Id: I4caa88f45a8304448f1b7fb475cc9d03480f6f98