[milo] console: fix builder link
Console impl assumed that builder id is a valid URL path of a builder. Use
model.BuilderIDLink. Also set BuilderSummary.ProjectID to the currently
served project.

This CL adds technical debt. Project must be a part of the builder ID, but that
is a larger refactoring or rewrite.

Bug: 800826
Change-Id: I60399ac779e3086fd5aa9cc779dd7e24cf3d5fa2
Reviewed-on: https://chromium-review.googlesource.com/860732
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Ryan Tseng <hinoka@chromium.org>