[cipd] General structure of CIPD backend plus some minor helpers.
It is pRPC server that exposes 2 services: cipd.Storage (for uploading files to
CAS) and cipd.Repository (for managing package repository).

It has 'backend' module that will be used to run package hash validation (using
instances of backend class). Tasks are routed with tq.Dispatcher instance.

Finally, for the duration of the development it is intended to be deployed on
top of Python implementation of the backend. 'gae.py' tool doesn't support this,
and it is simpler just to use 'gcloud app' directly.

R=iannucci@chromium.org, nodir@chromium.org
BUG=801423

Change-Id: I9893faf5cb1f05845cf5495c42073fc58daeb7ed