[vpython] Better Python argument handling.
Update the "-S" removal to work more correctly and robustly.

Update the "python" package's Python flag parsing to have insight into
the contents of the parsed flags, notably grouping each flag instance
with its value (if it has one). This results in a slightly cleaner flag
parsing routine.

Because flags are being deconstructed for potential manipulation
(notably removal of '-S'), update invocation sites to reconstruct flags
programmatically instead of using the raw argument vector. This
introduces the likelihood of variance between user input and Python
command, but if the deconstruction/construction is correct, this will
not have a user impact.

R=iannucci@chromium.org

Bug: 800662
Test: unit
Change-Id: Ibfac77e68508ea090d47d52f4e7b1b6a8479fbb4
Reviewed-on: https://chromium-review.googlesource.com/865975
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>