[vpython] Fix flag reassembly with separator.
Fix a bug where the "--" separator flag was considered a peer of other
interpreter flags on reassembly. It is critical that this flag occur
*after* all of the other flags, as is commands the interpreter to regard
all successors as positional arguments.

This was observed when a command was executed including a separator. The
initial parse added a "--" flag to the beginning of the Flags slice, and
isolation directives later appended other flags. Because "--"
maintained its position at the beginning, the isolation flags were
treated as positional arguments by the interpreter.

Bug: chromium:801960
Test: unit,reproduction
Change-Id: Ic8fab0eb2446dbd31caca54dee81b554615c3417
Reviewed-on: https://chromium-review.googlesource.com/867013
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>