[vpython] Do not SetIsolatedFlags when execing the actual script.
By default, we want vpython to behave as much like a real python
interpreter as possible, except that it's a python interpreter which
comes with whatever packages are specified in the correct .vpython spec.

To that end, this CL removes the -B option (turns off .py[co] writing), and
the -E option (ignore all PYTHON* envvars). PYTHONPATH is a user-critical
way of communicating with the python interpreter, so categorically
ignoring it is the wrong move. VPython will still clear it, if asked to
(with the VPYTHON_CLEAR_PYTHONPATH envvar), but now we'll actually respect
it otherwise.

This still will set the '-s' flag by default (to ignore the user site
directory), but provides an escape hatch (with VPYTHON_KEEP_USER_SITE) to
disable that.

R=dnj@chromium.org, nodir@chromium.org

Bug: 800662
Change-Id: Ic236d08a9aa285cb1af253af802fac01e2021ce4