Update patch set 2
Patch Set 2:

high level comment first: i am not sure about the design of feature. It adds a required field, but in practice the field will be useful only the first time or when table description changes (happens less often than other invocations of the tool). This is a lot of code for something that can be done manually very easily; note that the generated markdown does not have list of fields (this is why I was suggesting to provide a link to big query which has them, so that we don't have to write this code), so I don't see much value in automating this.

This CL is 264 LOC. Chances are, we will have <264 of lines in the generated markdown, and those lines are very simple (not code).

Patch-set: 2
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>