Update patch set 2
Patch Set 2:

Yeah, I see what you're saying. I agonized over that a bit when I started to write this. Is it really worth it? I wasn't sure either, but I landed on the side of yes, it is worth it. We don't do a great job of updating documentation as it is. If we rely on people to update the docs manually, and on reviewers to catch them, the documentation will slowly come out of date.

The markdown does include a link to the table, where you can see a list of fields.

Patch-set: 2
Reviewer: Katie Thomas <1167649@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>