Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Yeah, I see what you're saying. I agonized over that a bit when I started to write this. Is it really worth it? I wasn't sure either, but I landed on the side of yes, it is worth it. We don't do a great job of updating documentation as it is. If we rely on people to update the docs manually, and on reviewers to catch them, the documentation will slowly come out of date.
> 
> The markdown does include a link to the table, where you can see a list of fields.

How often do you expected the MD file to be updated? FWIU, only when a new table is created.

I don't think it is critical to have identical summary text in Markdown and table description in BQ. The description in MD could be just a oneliner giving an idea what a table is. If a reader is interested, they can click on the link and get a full description of the table with field schema and a link to the proto with even more comments. The purpose of the link one MD file is to have list of all tables, not necessarily accompanied with full descriptions.

If the MD file needs to be updated only once per table, I think developers can keep up with that frequency. It it not that often.

Patch-set: 2
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>