Update patch set 4
Patch Set 4:

I think "project" and "project path" concepts are not well defined and problematic. This code seems to assume a very specific file structure (main.go:127).

I'd rather prefer the table description to have an absolute Gitiles URL to the proto file. I think it would be more useful because a user can actually click on it and see the file contents, as opposed to wonder "what is project X? where is it?".

Please use git to figure out the googlesource host of the repo containing the proto (no new flags needed) and generate a link like
https://chromium.googlesource.com/infra/infra/+/master/go/src/infra/appengine/sheriff-o-matic/som/model/gen/events.proto

Bonus points if you make the link have absolute commit hash, e.g.
https://chromium.googlesource.com/infra/infra/+/90aebd1eefdd15ed2643d34a334a9b2c670507c4/go/src/infra/appengine/sheriff-o-matic/som/model/gen/events.proto

where the commit is the current commit, so that when a user clicks on proto link in the table description, they see the version of the proto this schema was generated from, as opposed to whatever happens to be current (which may be different if table schema wasn't updated since committing the proto)

Patch-set: 4
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>