Update patch set 2
Patch Set 2:

You know, the more I think about this approach the more I think it doesn't make sense. We've combined two entities which are not the same into one table just because they have a few columns in common. Suppose we had done this elsewhere: OSes and platforms have precisely the same columns (name, description), but they are in separate tables because they are separate entities. Similarly I believe physical hosts and VMs are separate entities that should be in separate tables.

Regarding entity differences, in the future, we will want to add physical properties to the machines, stuff like number of CPUs, RAM, disk space, etc. These belong to the machine. In the case of a physical host, like NICs, a physical host deployed on a machine owns the full number of CPUs, RAM, disk space, etc. so it's not necessary to specify these fields when creating a physical host. However, a VM may only be given access to a subset of cores, RAM, disk space, etc. Therefore it may become necessary in the future to add these columns to VMs. With this CL, this would mean adding more columns to the hosts table which must be NULL for physical hosts and non-NULL for VMs, creating further if conditions and further complicating the code.

This brings me to my second point. Having to check "if this", "if that" complicates reading, writing, and maintaining the code. By allowing two different kinds of host to use the same CreateHost API we create two branching codepaths. The original approach was a lot cleaner in this respect.

Lastly, we have the xor problem that MySQL cannot enforce that a host is physical xor virtual. This means we have to rely on the code to ensure that two cases don't occur: host is neither physical nor virtual, and host is both physical and virtual.

I've been thinking about another schema to address entity differences, code complexity, and the xor problem. Suppose we create the following tables:

hostnames (id, name, vlan_id)
- PRIMARY KEY (id)
- FOREIGN KEY (vlan_id) REFERENCES vlans (id)
- UNIQUE (name, vlan_id)

physical_hosts (id, hostname_id, machine_id, os_id)
- PRIMARY KEY (id)
- FOREIGN KEY (hostname_id) REFERENCES hostnames (id)
- FOREIGN KEY (machine_id) REFERENCES machines (id)
- FOREIGN KEY (os_id) REFERENCES oses (id)
- UNIQUE (hostname_id)
- UNIQUE (machine_id)

vms (id, hostname_id, physical_host_id, os_id)
- PRIMARY KEY (id)
- FOREIGN KEY (hostname_id) REFERENCES hostnames (id)
- FOREIGN KEY (physical_host_id) REFERENCES physical_hosts (id)
- FOREIGN KEY (os_id) REFERENCES oses (id)
- UNIQUE (hostname_id)

In these tables, none of the foreign keys will be nullable. The tables will be used as follows:

CreatePhysicalHost(r CreatePhysicalHostRequest):
- hostnameId := INSERT INTO hostnames(name, vlan_id) VALUES (r.Name, r.Vlan)
- physicalHostId := INSERT INTO physical_hosts(hostname_id, machine_id, os_id) VALUES (hostnameId, (SELECT id FROM machines WHERE name = r.Machine), (SELECT id FROM oses WHERE name = r.Os))

CreateVM(r CreateVMRequest):
- hostnameId := INSERT INTO hostnames(name, vlan_id) VALUES (r.Name, r.Vlan)
- INSERT INTO vms(hostname_id, vm_slot_id, os_id) VALUES (hostnameId, (SELECT p.id FROM physical_hosts p, hostnames h WHERE h.name = r.Name AND h.vlan_id = r.Vlan AND p.hostname_id = h.id), (SELECT id FROM oses WHERE name = r.Os))

Let's look at the desired properties you listed on https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/877379.

>host name is unique within VLAN
Derived from (name, vlan_id) uniqueness constraint in hostnames.

>a host is either bare metal or virtual, not both, not none
Derived from having two different tables for physical hosts and VMs.

>one bare metal host per machine
Derived from machine_id uniqueness constraint in physical_hosts.

>for a VM to exist, a bare metal host on the same machine must exit
Implied by physical_host_id in vms. Each VM has a non-NULL reference to a specific physical host, which has a non-NULL reference to a specific machine.

>VM's machine equals machine of the backing bare metal host
Implied by physical_host_id as above.

>VM's VLAN equals VLAN of the backing bare metal host
This is the only constraint we don't have in the database, however due to the way the code is structured we will have this anyways. When issuing a request to create a VM, users would specify the VM hostname, the physical hostname, and the VLAN ID the physical hostname exists on. With only one VLAN ID, we use that VLAN to both create the VM's hostname entry and look up the physical host the VM is running on.

With this schema I believe that we separate logically distinct entities, simplify branching code paths, and satisfy all but one constraint using the database (this CL satisfies all but two).

What do you think?

Patch-set: 2
Reviewer: smut <1003211@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0